Design operations

+++ app/Http/Controllers/DesignController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\CreateDesignRequest;
 use App\Http\Requests\UpdateDesignRequest;
 use App\Models\Category;
 use App\Models\Design;
+use App\Models\Product;
 use App\Repositories\DesignRepository;
 use App\Http\Controllers\AppBaseController as InfyOmBaseController;
 use Illuminate\Http\Request;
@@ -118,14 +119,18 @@ class DesignController extends InfyOmBaseController
     public function edit($id)
     {
         $design = $this->designRepository->findWithoutFail($id);
-
+        $mainCategory = Category::where('parent_id', 0)->whereNotIn('id',[0])->get()->toArray();
+        $subCategory = Category::whereNotIn('parent_id', array_keys($mainCategory))->whereNotIn('id',[0])->get()->toArray();
         if (empty($design)) {
             Flash::error('Design not found');
 
             return redirect(route('designs.index'));
         }
 
-        return view('designs.edit')->with('design', $design);
+        return view('designs.edit')
+            ->with('design', $design)
+            ->withMainCategory($mainCategory)
+            ->withSubCategory($subCategory);
     }
 
     /**
@@ -168,6 +173,12 @@ class DesignController extends InfyOmBaseController
             Flash::error('Design not found');
 
             return redirect(route('designs.index'));
+        }
+
+        $product = Product::where('code',$id)->get();
+        if(count($product)>0){
+            Flash::error('Design cannot be deleted as it is linked to Product.');
+            return redirect('/designs');
         }
 
         $this->designRepository->delete($id);
